feat(pug): move to pug-2.0.0-beta6
This is a quick and dirty rewrite of the library to use pug instead of jade. There needs to be a lot

more cleanup of unused code and unneeded options. More tests need to be introduced, potentially the

entire pug test suite for php. The runtime also needs to be cleaned up and tested more thoroughly.